// src/data/collections-name.enum.ts
export enum ECollectionsName {
    ADDRESS = 'Address',
    CATEGORY = 'Category',
    CONTACT = 'Contact',
    ITEM_ORDER = 'ItemOrder',
    LINK = 'Link',
    LIST_LINK = 'ListLink',
    LIST_PARTNER = 'ListPartner',
    LIST_PRODUCT = 'ListProduct',
    LIST_SLIDE = 'ListSlide',
    NOTIFICATION = 'Notification',
    ORDER = 'Order',
    PARTNER = 'Partner',
    PRODUCT = 'Product',
    REVIEW = 'Review',
    SLIDE = 'Slide',
    TOKEN_INVALID = 'TokenInvalid',
    USER = 'User'
}